implement personal rpc service
